A-Packages: Move classes into package(s) and fix imports
