+++ src/main/java/clover/Clover.java
@@ -1,6 +1,5 @@
-import java.util.Scanner;
-import java.util.ArrayList;
-import java.util.List;
+import clover.Command;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
