+++ src/main/java/clover/DeleteCommand.java
