+++ src/main/java/clover/MarkCommand.java
@@ -1,3 +1,5 @@
+import clover.Command;
+
 class MarkCommand extends Command {
     private final String arg;
     private final boolean mark;
