+++ src/main/java/clover/Parser.java
@@ -1,3 +1,5 @@
+import clover.*;
+
 public class Parser {
     public static Command parse(String input) throws DukeException {
         String in = input == null ? "" : input.trim();
